requirements.txt
openai>=1.0.0
faiss-cpu==1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0

#pip install -r requirements.txt
